data collection validation on the mdifier, and validation on intersts added

+++ controller/dataCollectionController.js
@@ -1,5 +1,9 @@
 const DataCollection = require("../models/dataCollection");
 
+const ensureDataCollectionModifier = (dataCollection, userId) => {
+  return dataCollection.userId === userId;
+};
+
 const getAllDataCollections = async (req, res) => {
   try {
     const dataCollections = await DataCollection.findAll();
@@ -25,7 +29,14 @@ const getDataCollectionById = async (req, res) => {
 const createDataCollection = async (req, res) => {
   const { location, interests, description, date } = req.body;
   const user = req.user;
-  console.log(user + "inside datacollection");
+
+  if (interests.length < 1 || interests.length > 255) {
+    return res.status(409).json({
+      message: interest + "is Not a Valid interests (length must be: 1-255)",
+      body: req.body,
+    });
+  }
+
   userId = user.userId;
   try {
     const newDataCollection = await DataCollection.create({
@@ -49,6 +60,21 @@ const updateDataCollection = async (req, res) => {
     if (!dataCollection) {
       return res.status(404).json({ error: "Data collection not found" });
     }
+    if (!ensureDataCollectionModifier(dataCollection, req.user.userId)) {
+      console.log(
+        ensureDataCollectionModifier(dataCollection, req.user.userId)
+      );
+      return res
+        .status(403)
+        .json({ error: "You aren't the creator of the collection" });
+    }
+
+    if (interests.length < 1 || interests.length > 255) {
+      return res.status(409).json({
+        message: interest + "is Not a Valid interests (length must be: 1-255)",
+        body: req.body,
+      });
+    }
     await dataCollection.update(req.body);
     res.status(200).json(dataCollection);
   } catch (error) {
@@ -62,6 +88,11 @@ const deleteDataCollection = async (req, res) => {
     const dataCollection = await DataCollection.findByPk(id);
     if (!dataCollection) {
       return res.status(404).json({ error: "Data collection not found" });
+    }
+    if (!ensureDataCollectionModifier(dataCollection, req.user.userId)) {
+      return res
+        .status(403)
+        .json({ error: "You aren't the creator of the collection" });
     }
     await dataCollection.destroy();
     res.status(204).end();
